Moved common schema attributes to shared package
